Tighten error surfacing from review-pr feedback (#28)
Addresses findings from the integrated #28 review:

- Mid-session unsolicited * BYE now surfaces its reason on in-flight commands.
  ConnectionActor captures the BYE code/text and, when the connection then drops,
  fails pending commands with connectionClosed(IMAPServerResponse(.bye, ...))
  instead of a generic disconnected. Captured-then-applied-at-teardown so a BYE
  that precedes a LOGOUT completion is unaffected.
- executeWithReconnect now classifies/throws the reconnect failure rather than the
  stale original error when reconnection itself fails.
- disconnect() logs a failed LOGOUT at debug instead of silently swallowing it.

Adds a close-after-response hook to MockIMAPServer and tests for the mid-session
BYE and reconnect-failure paths. Extracts MockIMAPServer into its own file to keep
IMAPIntegrationTests under the file-length limit.
